[API] Remove core reexports.
The main reason to re-export core from here was to simulate the
standard library to ensure that the generated bindings are able
to find all required data types at their expected locations.

This is longer necessary as the bindings are now generated with
an additional flag to ensure that they directly use core instead
of std.